Support Rails SafeStrings in helper return values
See complate/complate-stream#34 . This is still no really great solution since it doesn't allow automatic conversion in "input" values (a.k.a. `assigns`).Converting these doesn't seem to be a good idea since there may be non scalar values (like Objects) in there too.

Also it doesn't work for helpers in the 'old way' too since the bundle has to export the `safe` function. Since the bundle is generated by the user in the classic approach this can not be seen as guaranteed.